# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: intelligence
    description: Vantage — conflict, in context.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /intelligence/analyze:
    post:
      operationId: analyzeArticle
      tags: [intelligence]
      summary: Analyze a conflict article (URL or text)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeArticleBody"
      responses:
        "200":
          description: Structured intelligence brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceBrief"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The article URL could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests from this client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: The model provider failed or returned an unusable brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Generation capacity or daily budget reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /briefs/{id}:
    get:
      operationId: getBrief
      tags: [intelligence]
      summary: Read a stored brief by its id
      description: Returns the brief exactly as it was first generated. Nothing is regenerated and no model call is made.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9_-]{10,32}$"
      responses:
        "200":
          description: The stored brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceBrief"
        "404":
          description: No brief with this id is stored (it never existed, or it has been removed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /intelligence/stream:
    post:
      operationId: streamBrief
      tags: [intelligence]
      summary: Generate a brief and report progress as it is built
      description: |
        Takes a topic, an article URL or article text and answers with a
        `text/event-stream`. Events, in order:

        - `stage` — `{ "stage": "retrieving" | "writing" | "locating" }`
        - `sources` — `{ "sources": Source[], "retrieval": RetrievalStatus[] }`, sent once retrieval finishes
        - `brief` — the finished IntelligenceBrief (the last event on success)
        - `error` — an ErrorResponse (the last event on failure after the stream has started)

        Invalid input, rate limiting and unreadable article links are refused
        before the stream starts, with the usual JSON ErrorResponse and status.
        A brief already cached is sent as a single `brief` event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamBriefBody"
      responses:
        "200":
          description: Progress events followed by the brief
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests from this client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /intelligence/explore:
    post:
      operationId: exploreConflict
      tags: [intelligence]
      summary: Explore a conflict by topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploreConflictBody"
      responses:
        "200":
          description: Conflict exploration brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceBrief"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The article URL could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests from this client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: The model provider failed or returned an unusable brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Generation capacity or daily budget reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    AnalyzeArticleBody:
      type: object
      properties:
        article:
          type: string
          description: Full text of the conflict news article
          minLength: 50
          maxLength: 20000
        url:
          type: string
          description: Public http(s) URL of a conflict news article to fetch
          maxLength: 2048
      description: Provide exactly one of article or url. The server rejects both and neither.
    ExploreConflictBody:
      type: object
      properties:
        topic:
          type: string
          description: Conflict topic to explore (e.g. "Gaza conflict", "Sudan civil war")
          minLength: 3
          maxLength: 240
      required: [topic]
    Location:
      type: object
      properties:
        city:
          type: string
        country:
          type: string
        region:
          type: string
          description: Broader geopolitical region (e.g. "Eastern Europe", "Middle East")
        lat:
          type: [number, "null"]
          description: Set by geocoding the city and country, never by the model. Null when the place could not be confirmed. Lat and lng are null together.
        lng:
          type: [number, "null"]
      required: [city, country, region, lat, lng]
    Perspective:
      type: object
      properties:
        actor:
          type: string
          description: Name of the actor or group holding this perspective
        alignment:
          type: string
          enum: [Western, Regional, State Media, Civil Society, Affected Population]
        framing:
          type: string
          description: How this actor frames the event (1-2 sentences)
        interests:
          type: string
          description: Underlying interest shaping this framing (1 sentence)
      required: [actor, alignment, framing, interests]
    RelatedEvent:
      type: object
      properties:
        date:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [strike, escalation, negotiation, humanitarian, political]
        place:
          type: [string, "null"]
          description: Where the event happened, as a place name (e.g. "Jeddah, Saudi Arabia"). Null when it has no single location.
        lat:
          type: [number, "null"]
          description: Set by geocoding the place, never by the model. Null when the place could not be confirmed. Lat and lng are null together.
        lng:
          type: [number, "null"]
        searchQuery:
          type: string
          description: Google News search query to verify this event
      required: [date, title, description, type, place, lat, lng, searchQuery]
    CasualtyData:
      type: object
      properties:
        description:
          type: string
        civilianImpact:
          type: string
        allSides:
          type: string
      required: [description, civilianImpact, allSides]
    Source:
      type: object
      description: A record created by retrieval or supplied by the reader. Every field is set by the server; the model never writes sources, URLs or dates.
      properties:
        id:
          type: string
          description: Stable within one brief, e.g. "S1". Claims refer to sources by this id.
        kind:
          type: string
          enum: [news, reference, article]
          description: news is retrieved reporting, reference is background (Wikipedia), article is the text or link the reader supplied.
        provider:
          type: string
          description: Where the record came from, e.g. Bing News search, BBC News RSS, Wikipedia, Reader.
        publisher:
          type: string
        title:
          type: string
        url:
          type: [string, "null"]
        publishedAt:
          type: [string, "null"]
          description: ISO 8601 publication time when the provider gave one, otherwise null.
        retrievedAt:
          type: string
          description: ISO 8601 time the server retrieved the record.
        language:
          type: [string, "null"]
        country:
          type: [string, "null"]
          description: Country of the outlet when the provider reports it.
        excerpt:
          type: [string, "null"]
          description: The publisher's own summary or the reference extract, when available.
      required: [id, kind, provider, publisher, title, url, publishedAt, retrievedAt, language, country, excerpt]
    Claim:
      type: object
      description: A statement tied to retrieved sources. The server drops any claim whose source ids do not resolve or whose figures are absent from the cited text.
      properties:
        text:
          type: string
        sourceIds:
          type: array
          items:
            type: string
          minItems: 1
      required: [text, sourceIds]
    CoverageComparison:
      type: object
      properties:
        agreements:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
        differences:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
      required: [agreements, differences]
    RetrievalStatus:
      type: object
      properties:
        provider:
          type: string
        status:
          type: string
          enum: [ok, empty, failed]
          description: empty means the provider answered with nothing relevant; failed means it could not be reached or timed out.
        count:
          type: integer
      required: [provider, status, count]
    IntelligenceBrief:
      type: object
      properties:
        id:
          type: string
          description: Unguessable identifier assigned by the server. GET /briefs/{id} returns this same brief for as long as it is stored.
        generatedAt:
          type: string
          description: ISO 8601 time the server generated this brief. Set once; unchanged when served from cache.
        inScope:
          type: boolean
          description: False when the input is not a conflict, crisis or geopolitical tension.
        headline:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        summary:
          type: string
        actors:
          type: array
          items:
            type: string
        perspectives:
          type: array
          items:
            $ref: "#/components/schemas/Perspective"
        relatedEvents:
          type: array
          items:
            $ref: "#/components/schemas/RelatedEvent"
        escalationRisk:
          type: string
          enum: [Low, Medium, High]
        escalationReason:
          type: string
        historicalContext:
          type: string
        affectedPopulation:
          type: string
        keyQuestion:
          type: string
        casualtyData:
          $ref: "#/components/schemas/CasualtyData"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/Source"
        keyFacts:
          type: array
          description: Statements supported by the cited sources. Empty when nothing retrieved supports a claim.
          items:
            $ref: "#/components/schemas/Claim"
        coverage:
          $ref: "#/components/schemas/CoverageComparison"
        retrieval:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalStatus"
      required:
        - id
        - generatedAt
        - inScope
        - headline
        - location
        - summary
        - actors
        - perspectives
        - relatedEvents
        - escalationRisk
        - escalationReason
        - historicalContext
        - affectedPopulation
        - keyQuestion
        - casualtyData
        - sources
        - keyFacts
        - coverage
        - retrieval
    StreamBriefBody:
      type: object
      description: Exactly one of topic, url or article.
      properties:
        topic:
          type: string
          minLength: 3
          maxLength: 240
        url:
          type: string
          maxLength: 2048
        article:
          type: string
          minLength: 50
          maxLength: 20000
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Stable machine-readable code, e.g. INVALID_INPUT, NOT_FOUND, RATE_LIMITED, FETCH_BLOCKED, FETCH_FAILED, PROVIDER_UNAVAILABLE, MODEL_OUTPUT_INVALID, OVERLOADED, INTERNAL.
        message:
          type: string
          description: Safe to show to the person using the site.
        requestId:
          type: string
          description: Matches the X-Request-Id header and the server log line.
      required: [error]