# lib/api-zod/src/generated/types/analyzeArticleBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Provide exactly one of article or url. The server rejects both and neither.
 */
export interface AnalyzeArticleBody {
  /**
     * Full text of the conflict news article
     * @minLength 50
     * @maxLength 20000
     */
  article?: string;
  /**
     * Public http(s) URL of a conflict news article to fetch
     * @maxLength 2048
     */
  url?: string;
}