# lib/api-zod/src/generated/types/claim.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * A statement tied to retrieved sources. The server drops any claim whose source ids do not resolve or whose figures are absent from the cited text.
 */
export interface Claim {
  text: string;
  /** @minItems 1 */
  sourceIds: string[];
}