# lib/api-zod/src/generated/types/coverageComparison.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Claim } from './claim';

export interface CoverageComparison {
  agreements: Claim[];
  differences: Claim[];
}