# lib/api-zod/src/generated/types/exploreConflictBody.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExploreConflictBody {
  /**
     * Conflict topic to explore (e.g. "Gaza conflict", "Sudan civil war")
     * @minLength 3
     * @maxLength 240
     */
  topic: string;
}