# lib/api-zod/src/generated/types/relatedEvent.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RelatedEventType } from './relatedEventType';

export interface RelatedEvent {
  date: string;
  title: string;
  description: string;
  type: RelatedEventType;
  /**
     * Where the event happened, as a place name (e.g. "Jeddah, Saudi Arabia"). Null when it has no single location.
     * @nullable
     */
  place: string | null;
  /**
     * Set by geocoding the place, never by the model. Null when the place could not be confirmed. Lat and lng are null together.
     * @nullable
     */
  lat: number | null;
  /** @nullable */
  lng: number | null;
  /** Google News search query to verify this event */
  searchQuery: string;
}