# lib/api-zod/src/generated/types/retrievalStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RetrievalStatusStatus } from './retrievalStatusStatus';

export interface RetrievalStatus {
  provider: string;
  /** empty means the provider answered with nothing relevant; failed means it could not be reached or timed out. */
  status: RetrievalStatusStatus;
  count: number;
}